feat: 모든 action에 validation 추가
월드컵, 후보 추가 수정 삭제시 월드컵을 만든 user인지 확인하는 로직 모든
action에 추가함

+++ app/lib/actions/candidates/delete.ts
@@ -2,9 +2,18 @@
 
 import { pool } from '@/app/lib/db';
 import { revalidatePath } from 'next/cache';
+import { getSession } from '../session';
+import { validateWorldcupOwnership } from '../auth/worldcup-ownership';
 
-export async function deleteCandidate(worldcupId: string, candidateId: string) {
+export async function deleteCandidate(candidateId: string, worldcupId: string) {
   try {
+    const session = await getSession();
+    if (!session?.userId) {
+      throw new Error('로그인을 해주세요.');
+    }
+
+    await validateWorldcupOwnership(worldcupId, session.userId);
+
     const [result, fields] = await pool.query(
       `
         DELETE FROM candidate
@@ -15,6 +24,7 @@ export async function deleteCandidate(worldcupId: string, candidateId: string) {
     console.log(result);
   } catch (error) {
     console.log(error);
+    throw new Error('후보 삭제 실패...');
   }
   revalidatePath(`/worldcups/${worldcupId}/update-candidates`);
 }
